fix pycodestyle
