--- server/static/index.html
@@ -1,23 +1,0 @@
-
-<html>
-    <head>
-        <script src="//js.leapmotion/2.0.1/leap.min.js"></script>
-    </head>
-    <body>
-        <h1>hello 2018.meister</h1>
-        <div>
-            <p>index.html</p>
-        </div>
-        <script>
-            var controller=new Leap.Controller({enableGestures:ture});
-            controller.connect();
-            controller.on('frame',(frame)=>{
-                console.log(frame);
-            });
-            controller.on('deviceAttached',()=>{ console.log("deviceAttached"); });
-            controller.on('deviceConnected',()=>{ console.log("deviceConnected"); });
-            controller.on('deviceDisconnected',()=>{ console.log("devicsconnectede
-            controller.on('connect',()=>{ console.log("leaomotiion connected"); });
-        </script>
-    </body>
-</html>
